Remove unneeded stub in ClusterServerClient
